done with Meeting/visit site type functions
when user make a task and choose Meeting/visit site type small "is an offline meeting? " and Yes button will appear. if he leaves as it is then itss eitehr an online meeting or a site visit so the website textfield becomes required(is has url validator also :)).
now when he scrolls throw tasks the task that has Meeting/visit site and its not an offline meeting will display a small link icon button on bottom right when pressed it will appear a scrollable bottom sheet with website ready to launch direclt from my beautiful app